Bug #9337 fixed
git-svn-id: http://svn.berlios.de/svnroot/repos/phpfit/trunk@79 1a63dafe-9f17-0410-9df7-a0250be266ed

+++ eg/WeeklyTimesheet.php
@@ -26,7 +26,7 @@ class WeeklyTimesheet {
 			$extra = $this->standardHours - 40;
 			$tempHours = 40;
 		}
-		return $wage * ($tempHours + $extra * 1.5 + $this->holidayHours * 2);	
+		return (int)($wage * ($tempHours + $extra * 1.5 + $this->holidayHours * 2));	
         
 	}
